+++ Src/Private/Get-AbrVbrBackupServerInfo.ps1
@@ -45,7 +45,7 @@ function Get-AbrVbrBackupServerInfo {
                             if ($CimSession) {
                                 try { $DomainJoined = Get-CimInstance -Class Win32_ComputerSystem -Property PartOfDomain -CimSession $CimSession } catch { 'Unknown' }
                             }
-                                Write-PScriboMessage "Collecting Backup Server information from $($BackupServer.Name)."
+                            Write-PScriboMessage "Collecting Backup Server information from $($BackupServer.Name)."
 
                             if ($PssSession) {
                                 try {
@@ -72,19 +72,24 @@ function Get-AbrVbrBackupServerInfo {
                                     0 { "--" }
                                     default { $VeeamVersion.DisplayVersion }
                                 }
-                                'Database Server' = Switch ([string]::IsNullOrEmpty($VeeamDBInfo.SqlServerName)) {
+                                'Database Type' = Switch ([string]::IsNullOrEmpty($VeeamDBFlavor.SqlActiveConfiguration)) {
                                     $true { "--" }
-                                    $false { $VeeamDBInfo.SqlServerName }
-                                    default { 'Unknown' }
-                                }
-                                'Database Instance' = Switch ([string]::IsNullOrEmpty($VeeamDBInfo.SqlInstanceName)) {
-                                    $true { "--" }
-                                    $false { $VeeamDBInfo.SqlInstanceName }
+                                    $false { $VeeamDBFlavor.SqlActiveConfiguration }
                                     default { 'Unknown' }
                                 }
                                 'Database Name' = Switch ([string]::IsNullOrEmpty($VeeamDBInfo.SqlDatabaseName)) {
                                     $true { "--" }
                                     $false { $VeeamDBInfo.SqlDatabaseName }
+                                    default { 'Unknown' }
+                                }
+                                'Database Server' = Switch ([string]::IsNullOrEmpty($VeeamDBInfo.SqlHostName)) {
+                                    $true { "--" }
+                                    $false { $VeeamDBInfo.SqlHostName }
+                                    default { 'Unknown' }
+                                }
+                                'Database Port' = Switch ([string]::IsNullOrEmpty($VeeamDBInfo.SqlHostPort)) {
+                                    $true { "--" }
+                                    $false { "$($VeeamDBInfo.SqlHostPort)/TCP" }
                                     default { 'Unknown' }
                                 }
                                 'Connection Ports' = Switch (($VeeamInfo.BackupServerPort).count) {